Add blbuild and blflash scripts for updating bootloader
blbuild takes two arguments, the current x86 partition size on the
Arduino 101, and the desired size. The available choices are 144 (original)
216 and 256. It then builds the firmware once at the old size to get
a bootupdater.bin that will look in the current ARC location to find the
bootloader payload. Then it builds the firmware at the new size to get
the new bootloader. It then flashes the new bootloader to the old ARC
partition and the old updater to the X86 partition, and reboots. The
updater copies the bootloader from the ARC partition to the boot partition.

Then as long as you update your Zephyr to match the new ARC location, you
can flash both sides and all will be well. The X86 side will always work
because we haven't moved its start location at all. So the symptom of a
bootloader / Zephyr mismatch is that when you try to flash both X86 and
ARC, it doesn't work, but CONFIG_ARC_INIT=n will work.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>