Rename blbuild to blresize; leave a valid flashpack behind
Previously, the script patched up the flashpack in the a101fw tree after
it was created, so the zipfile there wasn't the same thing. Now we patch
up the bootupdater.bin file *before* it's included so the flashpack is
valid and can be taken to repeat this operation later without having to
build, etc.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>