Use ZJS_PARTITION env var as device partition size
By default, jsrunner will still expect it's targeting a factory Arduino 101
with 144KB x86_app partition. But with ZJS_PARTITION set to 216 or 256 it will
expect the corresponding size.

Using these features requires that your deps/zephyr be on the 'zjs' branch
which contains a patch that provides new configurations for A101:
arduino_101_factory_256, etc.

Also enhance the zjs_env.sh script to take an argument to set this variable.
$ source zjs_env.sh 256
for example, will set up the 256KB partition as default, which is our current
standard target for development.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>