Initial ble implementation
This is an attempt to implement something similiar to bleno apis as describe in
https://github.com/sandeepmistry/bleno. This checkin only covers the "on" events
setup and the advertising part.  With this, you should be able to see the device
from your Chrome browser. What's missing are Service and Characterics. Also,
there were random crashes that should be fixed once we figure it out how to
queue up the callbacks and execute it from the mainloop.