Reimplement callback queue using threadsafe fifo
We can't allocate memory from an ISR, so the trick is to have the callback
struct already allocated at the time the callback is bound, and then if
needed fill it in when the interrupt occurs. If needed, free resources used
as arguments.

The one who queues the callback again provides a callback to run the function,
this is where the code that knows how to determine the arguments will be. So
far, the two examples have had no arguments so this function was really the
same simple call. The difference between this function and just doing when
the interrupt comes in, is that this runs in task context, not ISR context.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>