[deps] Fix deps scripts to work with external upstream repos
The repos.txt file now allows you to specify the branch to check out, so
if we're not in sync with the tips of JerryScript or Zephyr for a time,
we can make sure the build still works. For now, I just set them both to
check out master.

Note: To work with a 256KB x86 partition on Arduino 101 requires a patch
to Zephyr, and I haven't yet figured out how to handle that. I could check
in the patch here and apply it, or I could submit it to Zephyr upstream
but I'm not sure it makes sense for wider consumption at this point.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>