[BLE] Implements onSubscribe and updateValue callback
This patch will call onSubscribe callbacks that are registered
in each characteristic and pass it the updateValue callback
function which is used for sending notification updates to
to client.  Since for now are not monitoring any actual
subscription events from the driver side, it just calls
onSubscribe when setService is called.

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>