[jsrunner] Fix to work with new top-level Makefile
Now generates script.h to outdir/include. Cleans outdir when -c is passed.
Runs dfu-util on the built image in outdir.

Support for 144KB partition and JTAG flashing is still broken (due to
recent Makefile patch), but this is an improvement. We could probably do
that by making 144KB the default in the top-level Makefile, that way
passing BOARD= would be able to override partition settings.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>