[BLE] Hack to fix connection problem
There's a know bug in our current BLE where the client
will not be able to establish the full bluetooth connection,
after trying to connect to our device, they will get an
erro like "Service Id not found" or busy-wait for the
service and characterstics to reply.  There's another problem
also in the mainloop if we don't sleep long enough, the
temperature notify update will not work, this patch helps
temporary fix these issues until we track down what the
real problem is.

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>