[GPIO] Replace set_callback function with on()
Update three samples to use the new method. Also, add a new test:
samples/test/ButtonEnable.js to test out dynamically adding and removing
a callback at runtime.

Important: The samples don't work right with the task_sleep(100) in main.c
that we've added as a workaround for the WebBluetooth demo. So for now,
remove that line if you want to test other samples. We'll come up with a
way for all the samples to work soon.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>