[BLE] Fixed bluetooth connection bug
Apparently, we didn't allocate memory for the primary service
and characterstic identifier which was initialized using the Zephyr
macros, so the struct was only allocated in the stack and
freed later, but the memory was not overwritten with certain
timing when the BLE connection happens.  With this patch,
we do not need to sleep work around anymore and the conection
is much more stable.

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>