[BLE] Updated BLE sample
Rewrite BLE sample to be a simple BLE service that allows
the client to write some arbitrary values to be stored
on the device and be able to read it back what he wrote.
The client was tested using the "nRF Connect" app on Android,
and can be tested with other BLE apps.

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>