[IPM] Refactor IPM message protocols
The message protocol has to be extended to support all the new modules
like I2C and Grove LCD. There's a size limit on the IPM message channel
that if exceeded, it won't go through, so refactoring the protocol to
send a pointer of the message instead of the actual message to save
message space.  Also the IPM message struct uses a union to support
different type of IPM messages like AIO, I2C and Grove LCD etc.

Also implemented a message queue on the ARC side to queue up the message
to be process in the mainloop, since we can't allocate memory on the ARC
it's statically created a queue of up to 10 messages.

Also updated the AIO notification logic to only send notification updates
when it detects a value change, so it doesn't flood the IPM channel.

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>