[ashell] Added the development mode 'A shell'
This shell focus is on creating an interface to be able to run commands and
arbitrary JavaScript directly from the USB port of a Zephyr enabled device.

Although 'a shell' would look like a common Unix Bash, it is heavily limited
due the constraints of the micro-controller, only a few basic functions are supported.

It is totally optional and to enable just add 'DEV=ashell' to your make

Basic usage and documentation is included in docs/ashell.md

This terminal will create a modem interface, usually at /dev/ttyACM0
using the USB port of the A101

This is how you will normally access this terminal:
`watch screen /dev/ttyACM0 115200`

Added a warning and the second shell.

The different shells run different commands and use different interfaces.

Signed-off-by: Sergio Martinez <sergio.martinez.rodriguez@intel.com>