[I2C] Adding string and array to Buffer
This adds the ability to create a Buffer with either
a string or an array of uint8_t.  It also implements
the 'write' method which allows you to copy a string
into the buffer you have already created.  Finally,
I updated the I2C.js demo to take advantage of these
new features.