[BLE] Updated HeartRateDemo.js to support multiple services
The original HeartRateDemo register 2 primary services, one
Heart Rate and one Generic Access.  This patch adds the
Generic Access service, but we have to increase our heap size
to 5k so the underlying ble can allocate enough structs to
store GATT attributes, or we would be out of memory and segfault.

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>