[AIO] Fix AIO API to support close() and drop abort()
Currently both close() and abort() does nothing, this patch
adds support for the close() api, so if there are any
registered callbacks for "change" events, they will no longer
be called.

Also dropping abort() as it's not needed since we don't return
promise in our current implementation, so we should wait till
the AIO api is finalized before we add it back.

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>