[BLE] Implement updateRssi() and handle rssiUpdate events
This is a initial implementation of the updateRssi() api
in BLE where each call to updateRssi() will trigger
an 'rssiUpdate' event to be fired, and application
can monitor the event by setting a on() callback handler.
Since there's no easy to get the RSSI value of the connected
device through Zephyr's api now, we are returning a fixed
dBm value for RSSI for now.

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>