[ashell] Improved serial handling & machine to machine transactions
Added extra machine readable tokens so the editor is able to read and use the outputs from the command line.

Changes to support uploading and check of code from the browser

Added support for parameters so commands can add a single character paramenters like `-x` or `-xyz`

makefile - Added support for grove LCD
jerry-code Split parse and run so we can check if the syntax is correct.

Fixed bug on carriage return that was introducing a \0 value after printing with carriage return.

Working on the uart to save some memory during the transactions on machine to machine transactions.

Uart is still a bit unstable in slow devices like old atom nucs

Signed-off-by: Sergio Martinez <sergio.martinez.rodriguez@intel.com>