[Sensor] Initial implementation of the Generic Sensor API
Initial implementation for the W3C Generic Sensor API and add
support for the Accelerometer and Gyroscope sensors using
the BMI160 inertia sensor built-in on the Arduino 101,
the current implementation is polling from the ARC and channeling
the data back to X86.

Also adding two sample JS code testing both APIs.

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>