Add Travis CI configuration file to setup continous integration. (#320)
This is the first iteration of enabling CI for the project. Each PR
and push command will trigger a message to the Travis CI infrastructure
which will attempt to build the project for the x86 (default), arc, linux
and Freedom Board targets, and report the results in the PR comments and
with an email. The exact behavior depends on the configuration selected
in the Travis CI control panel.

Signed-off-by: Francesco Balestrieri <francesco.balestrieri@intel.com>