[memory] Free global javascript objects and allow multiple restart
This adds the necessary code to supply JerryScript with callbacks
to free memory when an object is garbage collected.  It also makes
the changes needed to ensure global JavaScript objects get GC'd.
Finally - it adds the code needed by ashell to be able to run
a ZJS JavaScript file multiple times without rebooting.

Signed-off-by: Brian J Jones <brian.j.jones@intel.com>