[performance] New module implementing performance counter API.
Based on recent W3C API, formally defined in https://www.w3.org/TR/hr-time/
and informally in
https://developer.mozilla.org/en-US/docs/Web/API/Performance/now . As the
last link suggests, real-world usage boils down to just:

    t = performance.now();

even though formal spec defines more properties for "performance" object.

As defined in the spec for a browser, "performance" is a property of
global window object and thus implicitly present into context of a JS
script. We however define it as a module, so usage pattern is:

    var performance = require("performance");
    t = performance.now();

The returned value is a floating-point number of milliseconds since
arbitrary point in time. This means that its absolute value is
meaningless and only difference between 2 values make sense (and
represent time duration between calls which produced these 2 values).
As the returned value is a floating-point number, it may have higher
resolution than a millisecond (using fractional part). The actual
resolution however depends on a particular Zephyr board, and some
boards provide resolution only of 10 milliseconds.

This module is intended for benchmarking, and in extreme cases, for
synchronous delays, etc. Supported both in Linux and Zephyr builds.
A testcase is included.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>