[buffer] Use prototype inheritance to implement Buffer functions
This should save lots of runtime JS memory when there are multiple
buffers instantiated because only the single prototype needs to contain
the mappings to native API functions. I'd like to measure how much this
saves but expect we should replicate it across all APIs.

Also, added a zjs_obj_add_functions (plural) API to simplify creating a
large batch of native API mappings for an object.

Also, reorder cleanup for ashell so that jerry_cleanup() is the final
step and should leave the JS memory in its pristine state.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>