[samples] Fix Spaceship2.js timer bug (#484)
Stop using the event.value because it's unreliable because of button
bounce issues. Also, double edge interrupts are better used when the
same action will be triggered either way. Here a better approach to
make sure the lasers get turned off eventually is to check the button
status repeatedly in a timer. This makes it more like Spaceship1.js
but at least the lasers will turn on immediately, rather than syncing
to 250ms timer, which is important when you're leading a TIE fighter.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>