[uart] Use prototype inheritance for UART
This required implementing a fancier prototyping hierarchy than for
other modules because UART has its own functions and is an event
emitter. So added an argument to zjs_make_event to let you pass in a
different prototype to use; its prototype will be set to the event
emitter prototype.

This could be generalized further by checking for an existing
prototype and walking up the chain, but we can add that when it's
actually needed at some point.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>