[util] Add helper functions to reduce bugs with JrS strings
We have to be pretty careful with JerryScript strings to avoid using
too much stack space for a big string, make sure strings are null
terminated, etc. It makes sense to handle these issues in a single
place. This patch adds two new functions for copying JrS strings to
a buffer on the stack or allocating one on the heap (which must be
freed).

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>