[travis] Disallow new calls to jerry_string_to_char_buffer
This is to force people to use the new zjs_copy_jstring and
zjs_alloc_from_jstring. These calls require a maximum size to avoid
overflowing the stack or missing a null terminator.

Also, improve trlite by clone submodules of submodules instead of having
to download them. And fix a minor bug in error code handling.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>