[error] Implement standard error types
These are defined by Zoltan's iot-js-api. It works to check them with
the .name property or instanceof, the latter due to lots of painful
experimentation so you better use it!

Also, update native_require_handler to use the new standard error
macros and test these along with error constructors in a new
test-error.js file.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>