[sensor] Reduce Accelerometer and Gyroscope sampling frequency
This patch updates the sensor module allow adjusting the frequency
when creating Accelerometer and Gyroscope sensors, and set a default
of 20Hz if frequency is not defined.

Also fixed a issue where it should only send a change event
if the value has changed beyond the desire threshold, this
will significantly reduce the rate of events, since the ring buffer
and IPM channel cannot handle sensor events firing too rapidly.

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>