[build] Adjust x86 partition size without impacting bootloader
This was a big realization: Once Zephyr boots, it doesn't matter where
the bootloader thinks the ARC partition starts. Zephyr on X86 is in
charge of bringing up the ARC core. So by splicing images after they are
built, and flashing the right bits to where the bootloader thinks they
should go, we can change the partition without the bootloader being the
wiser. This removes the need for our 256KB flashpack tricks.

The magic is done with three simple dd commands after the A101 zephyr
build target completes.

This is controlled with a new SIZE argument to make, which defaults to
144 but can range from 144-296.

Also, add tests to the Makefile to sanity check the image sizes. Also,
make arc image get built and flashed along with the x86 image. The user
shouldn't really have to use "make arc" anymore, just "make" or "make
zephyr" and "make dfu" will now flash both images. The "ARC" image may
be a lot bigger because it's really the ARC image plus some portion of
the x86 image.

Update the "make help" command to reflect the removed targets and new
SIZE option.

Remove partition size code from zjs-env.h and prj.conf fragments.

Update READMEs to reflect the new state of affairs.

Also, remove jsrunner which is no longer used and referred to some of
this partition stuff.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>