[sensor] Fixed random hang after when frequency is set too high
This fixes a nasty bug that in the sensor module, it was calling
a jerry_function in the IPM callback which is a interrupt callback,
and it will cause the app to hang randomly when the frequency is
set at a high rate like more than 20Hz.

The lesson here is that we should really make sure we are not calling
into JerryScript engine directly or indirectly through one of our
helper functions inside a interrupt such as IPM message handlers.

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>