[sensor] Refactor sensor callbacks
Intead of shuffing the union of sensor reading to the callbacks,
only copy the the bytes needed for each type of sensor, for example,
Accelerometer and Gyroscope with X,Y,Z values will send 3 doubles,
where as Ambient Light will only send a single double.

Also rebased to use the new Zephyr api.

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>