[timers] Ensure coordinated timers execute in the order created
This changes the output of samples/Timers.js to match what happens in
Chrome and appears to be required by specification. We were previously
inserting the timers in reverse and this was visible to the users if they
created multiple timers with the same timeout.

Also found we weren't checking malloc for success when allocating the
timer callback argument storage, fixed that.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>