[callbacks] Retain removed callbacks until pending callbacks flushed
When a callback is removed, it may already have been signaled, which
means items are waiting in the ring buffer to be executed. If the same
callback ID were reassigned in the meantime, that pending callback
could execute the wrong code with the wrong arguments. So instead keep
the callback around, marked for removal, add a special flush command
into the ring buffer, and once it comes through we know it's safe to
actually remove the callback.

Also, reorder items in zjs_callback_t struct for better packing.

Fixes #682.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>