[util] Add macros/functions to standardize argument validation
The new ZJS_VALIDATE_ARGS macro and related macros are to be used at
the top of any JerryScript C binding to verify the quantity and types
of arguments, and return standard exceptions if there is a problem.

The comments above the macros in zjs_util.h explain the details of how
they are to be used.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>