[sensor] Fix broken sensor in ashell
The problem is that there's a conflict of the SPI bus in ashell
and sensor, so we need to explicitly disable SPI_0 and SPI_1 for
ARC when running in ashell.  We also need to increase the callstack
size for ashell, also have to decrease frequency to 20hz since Ashell is
so slow that even 100hz will be flooding the ringbuffer.

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>