[scripts] Enable debug mode when building ADC with SDK 0.9 (#836)
There is an issue with ADC when SDK version 0.9 is used and that is
causing a hang during the device boot (Zephyr issue- ZEP-1882). The
root cause seems to be some compiler optimization in GCC version
used in the SDK v0.9. So, this patch adds a workaround that enables
debug mode when SDK v0.9 is used and ADC is included in build.

Signed-off-by: Sudarsana Nagineni <sudarsana.nagineni@intel.com>