[general] Radically change our memory situation (Facepalm!)
Unbelievably, it turns out that newlib has a malloc that works way better
than Zephyr's k_malloc, and it's been right here all along. Sergio knew;
he was using it in ashell. We've been restricting ourselves to a 5120-byte
heap that allows only 63 max blocks to be allocated. And there's been >22K
available to malloc all along on Arduino 101!

Also, the overhead of the malloc implementation per block seems to be 8
bytes, while Zephyr's k_malloc overhead is 16 bytes.

So you can now allocate, for example, over 1000 16-byte blocks, over 100
256-byte blocks, or even multiple 4K blocks, and any size in between.

With the RAM= flag which can be set as high as 73 on A101, it goes even
higher automatically.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>