[trlite] Add support for running tests in parallel
You run trlite -j (but with no argument for now) to have it execute all
four "VMs" in parallel processes. It runs them silently but prints out
the last few lines if there are errors, and points you to the logs for
those that failed. (They are in /tmp/trlite-log*.)

To do this, it creates four separate tree clone directories, .trlite1
through .trlite4 in your ZJS_BASE. It will leave the ones there that had
a failure in case you need to debug further.

To support the parallel building, I had to fix a problem in analyze.sh
and Makefile where it was using a constant temp file name /tmp/zjs.js
which would then clash across the processes.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>