[mac] Add mac to Travis CI and fix build issues
This adds a BOARD=linux job on Travis for Mac OSX builds, it
includes fixes to all the warnings on mac as well as updated
scripts that will work on OSX.   It will take too much time
to build Zephyr targets on Mac OSX as it needs to fetch crosstool-ng
which is a cross compiler and then build it and fetches the arduino
ARC compiler to build the arc.

Also updating the Mac OSX build instructions to the latest.

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>