[build] Maintain both snapshot and parser versions of JrS library
Instead of replacing the same libjerry-core.a with a different version
depending on snapshot vs. parser build, maintain both with different
filenames. Then we don't have to rebuild JerryScript just because we
switched between those two types of builds, and don't need the snapshot
last_build file and related dangers of inconsistency.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>