[docs] Tweaks to text boxes w/ interactive commands for ashell/ocf
Originally, I meant to follow up with the same change as recently taken
for README.md, where we removed $ prompts to make it easier to copy and
paste the commands. But in both of these files, the prompt is more
important because it distinguishes bash commands from acm shell commands,
or from command output.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>