[headers] Fix header issues and checkheaders script; add to trlite
We'll see how this goes; the script gets mad if you use a symbol in a
header file without first including a header that defines it. This may
work because you have already included that dependency above in the C
file where this header is used, but it's not a great practice.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>