[build] ROM=144 default makes more sense to minimize flash times
If you don't like this, you can export ROM=256 or whatever you like in
your shell. But whenever you test small samples that can fit in the
default 144 it makes more sense to do that because it will flash
faster (and also wear out the flash on the device less quickly).

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>