[gpio] Add mock support to GPIO module
When built in mock mode, the gpio module exposes a 'mock' property
with the value 'True' and a 'wire' method that lets you pass two
pins to simulate connecting them with a wire for automated testing.

Update test-gpio.js to check for mock module and simulate wire from
IO7 to IO8 instead of asking user to set it up. This test now passes
15/15 when run in mock mode.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>