[ble] Fixes disconnect callback not getting called
The issue is that the c callback is immediately removed
after the callback is signaled, and the main loop have not
serviced the callback yet.  We don't need to allocate new callbacks
for each connection anyway, instead, just re-use the same
connect and disconnect callbacks, and each time create the callback
before it was signaled, and then remove it after callback is called.

Fixes #1088

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>