[webusb] Define device interface GUID
This defines device interface GUID in the extended properties
descriptor so that the Windows generic driver WinUSB will be
loaded automatically when the device is connected to the host.
This makes the installation process simple on Windows. Also,
we do not need to install the driver manually anymore, so
removing Zadig utility instructions from the docs.

Signed-off-by: Sudarsana Nagineni <sudarsana.nagineni@intel.com>