[ocf] Work-around to fix OCF segfault
There's a possible race condition in JerrScript or ZJS that parameters
were getting freed when promises are resolved and at the same time
the same jerry value is passed to another function through trigger
event.  This is a work-around patch to fix this until the root cause
is found.

Fixes #1316

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>