[analyze] Add virtualdeps feature and use it for a net-l2 dependency
This feature lets a "virtual dependency" like "net-l2" be fulfilled by
any of a number of different modules. In this network case, it lets the
JavaScript application remain transport-agnostic but at build time you
can choose whether to build for Bluetooth or Ethernet, for example. This
will surely require further evolution as we support a bigger variety of
hardware.

Also adds support for a new external ENC28J60 Ethernet board with the
A101 via SPI.

Add various new JSON files to implement this for net-l2.

Add documentation to MODULES.md about the virtualdeps feature and to
README.md about how to use the new ENC28J60 support.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>