[net-config] Update docs to require setting IP address
If the net-config module is included, the user needs to either set a
static IP address or use DHCP; no default address will be set. We could
potentially change this, but this requirement makes for better example
code because it's more realistic.

Added setStaticIP call to a sample and a test that needed it.

Fixes #1413

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>