[websockets] Fix two bugs causing ws connection memleak
The first was a circular reference to the connection object by the
listener functions; these were bound to a scope that contained the
websocket connection object in WebSocketServer.js, so even though
the connection dropped to 0 refcount, it wouldn't be freed. Fixed
this by releasing all the listener functions when the socket is
closed, after emitting the close event.

The second was a problem in timers where the callbacks for one-time
setTimeout timers weren't getting removed. As a result, the timeout
callback wasn't freed which like above still held a weak reference to
the connection object through the bound closure scope.

Fixes #1461

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>