[net] Clean up TCPClient6.js sample
Move connection options to a struct near the top of the file, with
the IP addresses and ports used called out. Add a print statement at
the top to identify the sample.

Also, change the server address from the link local address of James's
machine to the normal 2001:db8::2 address we use.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>