[ble] Fixed not able to set BLE address
The bug is that once bt_enable() is called, it's impossible to set
the BLE address, so we have to move bt_enable() to be called after
all modules and the JS is loaded, this guarantees that any API call to
setBleAddress will be called before bt_enable().

Also update documentation to reflect this change.

Fixes #1400

Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>