[net] Fix three bugs causing net memleak and corruption
The first was a circular reference to the connection object by the
listener functions, just like the recent problem fixed in websockets.

The second was releasing a socket object that had not been acquired;
this meant that we might still be accessing a value freed by the GC
and seemed to be causing corruption.

The third was a problem in the util list macros where the remove
function didn't return "true" in the case where the removed item was
the only item in the list.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>