[net] Acquire server until closed; store connections in server
The server object was being freed immediately if JS lost its reference
to it; instead it should only be freed if explicitly closed and all its
connections are closed.

To track this properly, store open connections within the server handle
they are associated with, and store a global list of servers instead of
connections.

Fixes #1468

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>