[net] Check for NULL server_h pointer in client socket
In a recent patch that restructured connection tracking for server
connection sockets, I forgot that client sockets use the same code, so
for these cases server_h is NULL. Also, still needed to track the list
of client connections in a global for the timeout feature to work as
implemented.

Also, merge post_closed into release_close since it was the only
caller. And add some comments for future clarity.

Fixes #1491, as well as the bus fault revealed after #1483 was fixed

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>