[tests] Add server error handling to dhcp ws tests
The payload is too large but it wasn't clear because server errors
were being ignored. We still need to look into why the payload is too
large and whether it's a bug or this test should extend the max payload.

The tests seem to work if you increase the payload size.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>