[net] Handle early close and use no_server handle for client conns
Because we defer the handling of new connections, it's possible for
the new connection socket to be closed before we handle the first data
packet, or even before we get around to accepting the new connection.
The correct thing to do in this case is to still report the new
connection, and report the data received, but mark the socket as
closed so we don't allow writing to it. Then once the connection and
data have been report it, we issue a close event.

To do this, we register the server handle, not the socket handle,
with tcp_received, and then look up the socket based on the context
the packet came in on. So we also had to do that for the client
sockets, for which we added a "no_server" stub handle.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>