[general] Revisit usage of memcpy everywhere
For copying structs, direct assignment is preferable to memcpy because
it is easier to read and gives more freedom to the compiler to be smart.

Using memcpy with strlen() as the size is the same as strcpy but with
an extra function call.

And sprintf is preferable to multiple strcats and memcpys to construct
a complex string.

Also made a few other simplifications along the way.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>