[callbacks] Fix a memory leak in removed callbacks (#1571)
Currently when zjs_timers_cleanup gets called, not all of the memory
in the callbacks is getting freed. The ring buffer isn't being
freed before the javascript ends. This change forces all the
memory to be freed when zjs_remove_all_callbacks gets called.

The problem was a callback getting removed twice resulted in
a double free, which really means the object never gets freed.

Signed-off-by: Brian J Jones <brian.j.jones@intel.com>