[websockets] Fix crash if no subprotocols were included
If the node client did not specify any websocket subprotocols
to use, the server would crash if the acceptHandler was
defined in the server init object. This was due to the protocol
array not being initialized if no Sec-WebSocket-Protocol
tag was present in the accept packet.

This initially sets the protocol array to undefined, and if its
still undefined when the acceptHandler is called it initializes
it to an empty array.

Signed-off-by: James Prestwood <james.prestwood@intel.com>