[ashell] Fix for ashell boot
Currently ashell boot doesn't work with javascript with callbacks
or timers.  This is because once the boot.cfg JS gets called, it
goes into an infinite loop waiting for uart. Until it gets the
connection, it won't return.  This means the main loop doesn't get
control back, meaning callbacks are not processed.  This breaks
the code up a bit so that it checks if the connection is ready
when uart_process gets called, if not try to connect.  If the
connection fails, just return and go through the main loop again.
Continue until a connection is made.

Signed-off-by: Brian J Jones <brian.j.jones@intel.com>