[gpio] Rewrite GPIO input/output sample tests as board-agnostic
Previously, I had written K64F-specific tests to help quickly test
which GPIO pins worked as inputs and outputs. Now, to test the
Nucleo-F411RE board I've rewritten them to be able to work with
various boards. I still haven't added "board support" with friendly
pin mappings for this board, so the tests are written in terms of full
Zephyr pin names for this board for now.

Signed-off-by: Geoff Gustafson <geoff@linux.intel.com>